Remove pytest configuration and related test files
- Deleted pytest.ini and requirements.txt as part of the cleanup process.
- Removed all test files related to blockchain functionality, configuration management, DEX operations, input/output functionality, and main functions to streamline the project.
- This change simplifies the project structure and focuses on core functionalities without the overhead of testing frameworks.

+++ src/input.cpp
@@ -52,7 +52,7 @@ void create_config_file(int thread_count, FunctionStats& stats) {
 #else
         struct stat file_stat; // Structure for file stats
         stat("neozork-config", &file_stat); // Get file stats
-        stats.disk_usage_bytes = file_stat.st_size; // Update disk usage
+        stats.disk_usage_bytes = static_cast<size_t>(file_stat.st_size); // Update disk usage
 #endif
     }
